Move pipeline tests
