+++ src/encoded/tests/test_types_pipeline.py
@@ -11,24 +11,3 @@ def test_analysis_step_version_name_calcprop(testapp, analysis_step, analysis_st
     assert analysis_step_version['minor_version'] == 0
     assert analysis_step_version['name'] == 'fastqc-step-v-1-0'
     assert analysis_step_version['@id'] == '/analysis-step-versions/fastqc-step-v-1-0/'
-
-
-@pytest.mark.parametrize(
-    'status',
-    [
-        'released',
-        'in progress',
-        'archived',
-        'deleted',
-        'revoked',
-    ]
-)
-def test_pipeline_statuses(status, testapp, pipeline):
-    testapp.patch_json(pipeline['@id'], {'status': status})
-    res = testapp.get(pipeline['@id'] + '@@embedded').json
-    assert res['status'] == status
-
-
-def test_pipeline_invalid_statuses(testapp, pipeline):
-    with pytest.raises(Exception):
-        testapp.patch_json(pipeline['@id'], {'status': 'active'})
